+++ workbench/experiments/testScript.py
@@ -1,5 +1,6 @@
 from mido import MidiFile
 from utils.midi_funcs import *
+from utils.time_series_funcs import *
 import numpy as np
 from itertools import groupby
 
@@ -11,7 +12,14 @@ track = midifile.tracks[0]
 # print(tempo)
 #
 parsed = simple_parse_midi(midifile, 0, get_tempo(midifile))
-# print(parsed)
+print(parsed[0])
+
+features_to_midi(parsed, get_tempo(midifile), midifile.ticks_per_beat, 'midi/data/resynth/fromParsed.mid')
+
+print("a"+" "+"space")
+
+# print (abs(-1))
+
 #
 # sig = get_key(mid)
 # print(sig)
@@ -27,26 +35,26 @@ parsed = simple_parse_midi(midifile, 0, get_tempo(midifile))
 # thisFile.write(features)
 # thisFile.close
 # print(parsed)
-
-make_time_series(midifile, 0, "./time_series")
-
-time_series = read_time_series("./time_series") # read the time series into memory
-time_signature = {'numerator': time_series.get('numerator'), \
-  'denominator': time_series.get('denominator'), \
-  'clocks_per_click': time_series.get('clocks_per_click'), \
-  'notated_32nd_notes_per_beat': time_series.get('notated_32nd_notes_per_beat'), \
-}
-tempo = time_series.get('tempo')
-ticks_per_beat = time_series.get('ticks_per_beat')
-time_series = time_series.get('time_series')
-
-# time_signature = get_time_signature(midifile)
-time_series_to_midifile(time_series, tempo*2.5, ticks_per_beat, time_signature, "./midi/data/resynth/testMIDI.mid")
-
-# notes = extract_notes_from_time_series(time_series)
-# rhythm = extract_rhythm_from_time_series(time_series)
-
-# time_series_file = open("./time_series", "r") # read the file in
-# time_series = time_series_file.readline() # read in the file; it's just one line
-# time_series = list(map(int, time_series.split(" "))) # split and convert to integers
-# print (rhythm)
+#
+# make_time_series(midifile, 0, "./time_series")
+#
+# time_series = read_time_series("./time_series") # read the time series into memory
+# time_signature = {'numerator': time_series.get('numerator'), \
+#   'denominator': time_series.get('denominator'), \
+#   'clocks_per_click': time_series.get('clocks_per_click'), \
+#   'notated_32nd_notes_per_beat': time_series.get('notated_32nd_notes_per_beat'), \
+# }
+# tempo = time_series.get('tempo')
+# ticks_per_beat = time_series.get('ticks_per_beat')
+# time_series = time_series.get('time_series')
+#
+# # time_signature = get_time_signature(midifile)
+# time_series_to_midifile(time_series, tempo*2.5, ticks_per_beat, time_signature, "./midi/data/resynth/testMIDI.mid")
+#
+# # notes = extract_notes_from_time_series(time_series)
+# # rhythm = extract_rhythm_from_time_series(time_series)
+#
+# # time_series_file = open("./time_series", "r") # read the file in
+# # time_series = time_series_file.readline() # read in the file; it's just one line
+# # time_series = list(map(int, time_series.split(" "))) # split and convert to integers
+# # print (rhythm)
